Remove setNext method by initialising an empty next frame when frame is created
SetNext on the final frame throws an exception which we thought was dubious.
We decided to pass in the nextframe in the constructor of Frame.
This means we have to redo the way we manage frames and create all the frames upfront.
We cannot add the next frame when looping from 0-9 since it has not been created yet.
eg: frames[i] = Frame(frames[i+1]) - frames[i+1] doesn't exist yet

so we loop backwards to add the next frame.