Remove final frame roll count
We use IRoll to keep the concept of a Roll intact while distinguishing it from NotRolled. E.g. A roll shouldn't say that it has been rolled and should be evident from the nature of a roll. We did this due to the null object pattern applied to NotRolled and also that it was a Roll.